+++ packages/solid-pacer/src/async-debouncer/createAsyncDebouncer.ts
@@ -1,5 +1,5 @@
 import { AsyncDebouncer } from '@tanstack/pacer/async-debouncer'
-import { useStore } from '@tanstack/solid-store'
+import { shallow, useSelector } from '@tanstack/solid-store'
 import { createEffect, onCleanup } from 'solid-js'
 import { useDefaultPacerOptions } from '../provider/PacerProvider'
 import type { Store } from '@tanstack/solid-store'
@@ -50,8 +50,8 @@ export interface SolidAsyncDebouncer<
   readonly state: Accessor<Readonly<TSelected>>
   /**
    * @deprecated Use `debouncer.state` instead of `debouncer.store.state` if you want to read reactive state.
-   * The state on the store object is not reactive, as it has not been wrapped in a `useStore` hook internally.
-   * Although, you can make the state reactive by using the `useStore` in your own usage.
+   * The state on the store object is not reactive, as it has not been wrapped in a `useSelector` hook internally.
+   * Although, you can make the state reactive by using the `useSelector` in your own usage.
    */
   readonly store: Store<Readonly<AsyncDebouncerState<TFn>>>
 }
@@ -194,14 +194,18 @@ export function createAsyncDebouncer<
     selector: (state: AsyncDebouncerState<TFn>) => TSelected
     children: ((state: Accessor<TSelected>) => JSX.Element) | JSX.Element
   }) {
-    const selected = useStore(asyncDebouncer.store, props.selector)
+    const selected = useSelector(asyncDebouncer.store, props.selector, {
+      compare: shallow,
+    })
 
     return typeof props.children === 'function'
       ? props.children(selected)
       : props.children
   }
 
-  const state = useStore(asyncDebouncer.store, selector)
+  const state = useSelector(asyncDebouncer.store, selector, {
+    compare: shallow,
+  })
 
   createEffect(() => {
     onCleanup(() => {